requirements.txt
twikit
dotenv
matplotlib
google-generativeai